Toggle theme integrated with DB.
