+++ frontend/src/components/ThemeToggle/index.jsx
@@ -4,6 +4,8 @@ import { ColorModeContext } from "../../App"
 import { useTheme } from '@mui/material/styles';
 import Switch from '@mui/material/Switch';
 import MoonIcon from '../../assets/icons/moon.svg'
+import { useGetOnePreference } from '../../graphql/getOnePreference';
+import { useUpdatePreferences } from '../../graphql/updatePreferences';
 
 const DarkToggle = styled(Switch)(({ theme }) => ({
     padding: 8,
@@ -70,8 +72,38 @@ const DarkToggle = styled(Switch)(({ theme }) => ({
   }));
 
 const ThemeToggle = (props) => {
+  const getOnePreference = useGetOnePreference()
+  const updatePreferences = useUpdatePreferences()
+
   const theme = useTheme();
   const colorMode = React.useContext(ColorModeContext);
+
+  // Retrieve color mode on load
+  React.useEffect(() => {
+    (async () => {
+      const colorModeResponse = await getOnePreference({preference: "theme"})
+
+      // If color mode in DB doesn't match current color mode, toggle.
+      if(!colorModeResponse.errors){
+        if (colorModeResponse.value !== theme.palette.mode) {
+          colorMode.toggleColorMode()
+        }
+      }
+    })();
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [])
+
+  // Toggle color mode on change
+  React.useEffect(() => {
+    (async () => {
+      const input = {input:{preference: "theme", value: theme.palette.mode} }
+      const updateEnvironmentsResponse = await updatePreferences(input);
+      if (updateEnvironmentsResponse.errors){
+        console.log('Unable to update color theme')
+      } 
+    })();
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [theme.palette.mode])
 
   return (
     <DarkToggle {...props}
